requirements.txt
python-binance
Flask